S-01024 2 metricas realizadas

+++ code/self_managment/src/main/java/com/self_managment/persistance/dao/SummaryDao.java
@@ -11,8 +11,8 @@ public interface SummaryDao extends GenericDao<Summary, Serializable> {
 	    Date dateFrom, Date dateTo);
 
     public Double getAverageTalkTime(Agent agent, Date dateFrom, Date dateTo);
-    public Double getNCH(Agent agent, Date dateFrom, Date dateTo);
-    public Double getTransferPCT(Agent agent, Date dateFrom, Date dateTo);
+    public Long getNCH(Agent agent, Date dateFrom, Date dateTo);
+    public Long getTransferPCT(Agent agent, Date dateFrom, Date dateTo);
     
     public Long getTotalLoggedTime(Agent agent, Date dateFrom, Date dateTo);
 }
